Style: 업데이트 프로퍼티 이동
- PhotoBoothUpdateProps를 entity -> Dto파일로 이동

Related to #4
